pyproject.toml
[project]
name = "SDR_analyser"
version = "0.0.1"
description = "Default template for PDM package"
authors = [
    {name = "Dylan Hunter Whittingham", email = "dylanwhittingham@outlook.com"},
]
dependencies = [
    "pyrtlsdr[lib]>=0.3.0",
    "setuptools>=69.1.1",
    "numpy>=1.26.4",
    "matplotlib>=3.8.3",
    "sigmf>=1.2.0",
    "autopep8>=2.1.0",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "GPL-3.0-only"}


[tool.pdm]
distribution = false
